vista detallecompra y funcionalidad del boton detalle compra

+++ src/main/java/co/com/tienda/dao/IOrdenDao.java
@@ -1,6 +1,7 @@
 package co.com.tienda.dao;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,7 +10,8 @@ import co.com.tienda.domain.Orden;
 import co.com.tienda.domain.Usuario;
 
 @Repository
-public interface IOrdenDao extends JpaRepository<Orden, Long>{
+public interface IOrdenDao extends JpaRepository<Orden, Integer>{
     
     List<Orden> findByUsuario(Usuario usuario);
+    //Optional<Orden> findById(Integer id);
 }
